+++ gDRutils/R/getMetaData.R
@@ -26,7 +26,7 @@ getMetaData <- function(data, discard_keys = NULL) {
   stopifnot(any(inherits(data, "data.frame"), inherits(data, "DataFrame")))
   checkmate::assert_character(discard_keys, null.ok = TRUE)
 
-  cell_id <- gDRutils::get_env_identifiers("cellline", simplify = TRUE)
+  cell_id <- get_env_identifiers("cellline", simplify = TRUE)
 
   data <- S4Vectors::DataFrame(data)
   all_data_cols <- colnames(data)
@@ -36,11 +36,11 @@ getMetaData <- function(data, discard_keys = NULL) {
     setdiff(
       all_data_cols,
       c(
-        gDRutils::get_header("raw_data"),
-        gDRutils::get_header("normalized_results"),
-        gDRutils::get_header("averaged_results"),
-        gDRutils::get_header("metrics_results"),
-        gDRutils::get_env_identifiers("well_position", simplify = TRUE),
+        get_header("raw_data"),
+        get_header("normalized_results"),
+        get_header("averaged_results"),
+        get_header("metrics_results"),
+        get_env_identifiers("well_position", simplify = TRUE),
         "Barcode",
         "Template",
         # not sure how to handle these ones ....    < --------
@@ -52,8 +52,8 @@ getMetaData <- function(data, discard_keys = NULL) {
 
   # Remove cell-related metadata.
   nocell_metavars <- setdiff(metavars,
-                             c(gDRutils::get_env_identifiers("cellline", simplify = TRUE), 
-                               gDRutils::get_header("add_clid")))
+                             c(get_env_identifiers("cellline", simplify = TRUE), 
+                               get_header("add_clid")))
   singleton_cols <- vapply(nocell_metavars,
                            function(x) {
                              nrow(unique(conditions[, x, drop = FALSE])) == 1L
@@ -62,16 +62,16 @@ getMetaData <- function(data, discard_keys = NULL) {
 
   # Remove drug metadata and duration.
   constant_metavars <- setdiff(nocell_metavars[singleton_cols],
-                               c(gDRutils::get_env_identifiers("drug", simplify = TRUE),
-                                 gDRutils::get_env_identifiers("drug_name", simplify = TRUE),
-                                 gDRutils::get_env_identifiers("duration", simplify = TRUE)
+                               c(get_env_identifiers("drug", simplify = TRUE),
+                                 get_env_identifiers("drug_name", simplify = TRUE),
+                                 get_env_identifiers("duration", simplify = TRUE)
                                  ))
 
-  unique_metavars <- c(intersect(c(gDRutils::get_env_identifiers("cellline", simplify = TRUE),
-                                   gDRutils::get_header("add_clid"),
-                                   gDRutils::get_env_identifiers("drug", simplify = TRUE),
-                                   gDRutils::get_env_identifiers("drug_name", simplify = TRUE),
-                                   gDRutils::get_env_identifiers("duration", simplify = TRUE)),
+  unique_metavars <- c(intersect(c(get_env_identifiers("cellline", simplify = TRUE),
+                                   get_header("add_clid"),
+                                   get_env_identifiers("drug", simplify = TRUE),
+                                   get_env_identifiers("drug_name", simplify = TRUE),
+                                   get_env_identifiers("duration", simplify = TRUE)),
                                  metavars),
                        nocell_metavars[!singleton_cols])
 
@@ -84,7 +84,7 @@ getMetaData <- function(data, discard_keys = NULL) {
     }
 
   pattern <- sprintf("^%s*|^%s*|^%s$",
-                     gDRutils::get_env_identifiers(c("drug", "drug_name", "duration"), simplify = FALSE))
+                     get_env_identifiers(c("drug", "drug_name", "duration"), simplify = FALSE))
   cl_entries <- cl_entries[!grepl(pattern, cl_entries)]
 
   ## colData