+++ gDRutils/inst/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Change to v.1.0.34
+- add helper function for MAE/experiments
+
 # Change to v.1.0.33
 - move constant fit warning